+++ actix-web/src/middleware/logger.rs
@@ -303,7 +303,7 @@ where
 
         ready(Ok(LoggerMiddleware {
             service,
-            inner: self.0.clone(),
+            inner: Rc::clone(&self.0),
         }))
     }
 }
@@ -657,11 +657,7 @@ impl FormatText {
 
             FormatText::ResponseHeader(ref name) => {
                 let s = if let Some(val) = res.headers().get(name) {
-                    if let Ok(s) = val.to_str() {
-                        s
-                    } else {
-                        "-"
-                    }
+                    val.to_str().unwrap_or("-")
                 } else {
                     "-"
                 };
@@ -705,11 +701,7 @@ impl FormatText {
             FormatText::RequestTime => *self = FormatText::Str(now.format(&Rfc3339).unwrap()),
             FormatText::RequestHeader(ref name) => {
                 let s = if let Some(val) = req.headers().get(name) {
-                    if let Ok(s) = val.to_str() {
-                        s
-                    } else {
-                        "-"
-                    }
+                    val.to_str().unwrap_or("-")
                 } else {
                     "-"
                 };
@@ -747,7 +739,7 @@ impl FormatText {
 /// Converter to get a String from something that writes to a Formatter.
 pub(crate) struct FormatDisplay<'a>(&'a dyn Fn(&mut fmt::Formatter<'_>) -> Result<(), fmt::Error>);
 
-impl<'a> fmt::Display for FormatDisplay<'a> {
+impl fmt::Display for FormatDisplay<'_> {
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         (self.0)(fmt)
     }